logi zebrane przez moduł servera wypychane sa przez WebSocket

+++ common/def.h
@@ -11,7 +11,7 @@
 namespace cm {
 
 typedef unsigned AutoId;
-typedef int TCPPort;
+typedef quint16 TCPPort;
 typedef QString ColumnName;
 typedef std::map<ColumnName, QVariant> Params;
 typedef QString Message;
